finish dex implementation
