+++ common/sso.go
@@ -1,0 +1,5 @@
+package common
+
+const (
+	SSOController = "sso-controller"
+)
